+++ src/main/java/com/project/shoppincart/service/ProductService.java
@@ -7,9 +7,9 @@ import java.util.Optional;
 
 public interface ProductService {
     Product save(Product product);
-    Optional<Product> getProductById(Integer id);
+    Optional<Product> getProductById(Long id);
     Product update (Product product);
-    void delete(Integer id);
+    void delete(Long id);
     List<Product> findAll();
     List<Product> saveList(List<Product> products);
 }
